requirements.txt
flask==3.0.2
psycopg2-binary
requests==2.31.0
bs4
chardet
pytz
requests==2.31.0
Gunicorn==20.1.0
instaloader==4.12
selenium==4.23.1
mysqlclient==2.2.4
sqlalchemy==2.0.32
flask_sqlalchemy
python-dotenv
instagram-private-api==1.6.0
google-api-python-client==2.141.0
tweepy==4.14.0
PyJWT==2.6.0
Flask-Migrate==4.0.4



